This is a classic example of the "imperative style" Cucumber anti-pattern. The scenario contains too much detail, so that you can't see what's actually being specified.
The rules are implicit, and impossible to guess.

It makes for poor documentation of the system behaviour.

Deliberate ambiguities / bugs:

- Are the credits deducted per shout or per listener? It's not clear,
  but the implementation is per shout.
- Does the keyword tax apply per shout or per word? It's not clear, but
  the implementation is per word.
- Do both rules apply on some messages e.g. long and with a keyword.
  It's not clear, but the implementation is that all rules apply.
- What happens when the shouter's credits run out? currently long
  messages will stop sending, but the keyword tax will just stop being
  applied. This is almost certainly a bug.

Unit tests:

We've had to put in a dirty hack here in the specs to get them to carry
on passing now that we have the credits functionality.

This hack is a classic symptom of test-last programming. If we'd TDD'd
this, we'd have noticed that the test was hard to write, listened to
that feedback, and reconsidered our design. Now, because we've already
written the messy code and this failing test is just an inconvenience,
our motivation is to just bash it into shape and move on.

So we get no design feedback (or we ignore it anyway) and we get a
messy hack in the test.